// packages/core/src/drivers/driver.interface.ts
export interface UniversalImageData {
    data: Uint8Array | Uint8ClampedArray;
    width: number;
    height: number;
}